pyproject.toml
[tool.poetry]
name = "duden"
version = "0.18.0"
description = "CLI-based german dictionary"
authors = ["Radomír Bosák <radomir.bosak@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/radomirbosak/duden"
keywords = ["duden", "duden.de", "dictionary", "cli", "word", "german"]
classifiers= [
    "Environment :: Console",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Natural Language :: German",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.4",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Topic :: Education",
]
packages = [{include = "duden"}]
include = ["duden/locale/**/*.*"]

[tool.poetry.dependencies]
python = "^3.7.2"
beautifulsoup4 = "^4.11.1"
pyxdg = "^0.28"
PyYAML = "^6.0"
requests = "^2.28.1"
crayons = "^0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pylint = "^2.15.2"
black = "^22.8.0"
isort = "^5.10.1"

[tool.poetry.scripts]
duden = 'duden.cli:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pylint.VARIABLES]
additional-builtins = ["_"]

[tool.pylint.'MESSAGES CONTROL']
disable = ["consider-using-f-string", "fixme"]
